Added in /pending an article about deployment with Apache and Nginx
This article shows the problems encountered in the deployment of
Dancer in a legacy environment, with PHP sites hosted on the same
machine and Apache running.

It turned out that the solution was to make nginx work as a proxy for
both Apache and Nginx.

FCGI setup is not commented nor discussed.