Withdraw 2013 calendar :(
Saddening, I really don't want to have to do this, but I'm just not going to
have the time to spend on the calendar this year, and haven't had contributed
articles - so I either spend a fair while still selecting and re-running
previous articles, or can it for this year and come back in force next year.

Sorry all.