Fixed misunderstood send_as() change in D2
The prior version of this article gave advice that only works in older
versions of D2 and in D1. Rewrote the discussion talking about how to
return the same data in both HTML and JSON forms accordingly.  Also
reworked the discussion of "parts," making it more implicit to the
discussion rather than tangential.