Clarify redirect_address works with all transports.
When I first read this, I thought "wait, you're talking about using the
Redirect transport, but your config example is using the Sendmail transport" -
on reading up a bit, I see that the plugin will first set up the transport using
the parameters given, then, if the `redirect_address` parameter was provided,
will pass that transport object to the Redirect transport's constructor and use
that as the transport.

So, I think this clarifies it a little.